+++ NEWS.md
@@ -1,3 +1,6 @@
+# VIM 6.x.x
+ - fix infinite loop in matchImpute in case all observations of a variable are missing
+
 # VIM 6.2.3
 - default robust regression method for irmi for numeric variables changes from rlm to lmrob.